deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
pillow
scikit-learn == 1.4.2